kernel/block_cache: Fix memory leak in sub-transaction handling.
This code has been broken like this since 2012, so I am skeptical
that it is the cause of the more recently observed #17463.
It is however possible that more recent changes caused things
to get noticeably worse for some reason.

This may, however, resolve some of the other KDLs, e.g. #12847.