freebsd_network: Automatically switch to jumbo-sized MTUs if possible.
Unfortunately there is no way to ask if a device supports jumbo MTUs
besides actually activating it. But that's fine, because for now,
Haiku's own network stack does not actually support changing MTUs
at the device level! Probably that should be fixed, too, but that
is a problem for another day.

With preceding commits, this fixes #17728, and all BSD network drivers
will use jumbo frames if possible by default.